requirements.txt
# Основные зависимости
aiogram>=3.3.0
python-dotenv>=1.0.0
APScheduler>=3.10.1

# Для работы с данными
pandas>=2.0.0
openpyxl>=3.1.2

# Для кэширования
cachetools>=5.3.0

# Для работы с датами
python-dateutil>=2.8.2

# Для асинхронной работы с файлами
aiofiles>=23.1.0

# Для работы с SQLite
aiosqlite>=0.19.0

# Для типизации
typing-extensions>=4.5.0